\name{acebayes-internal}
\alias{lrlin}
\alias{lrquad}
\alias{colin}
\alias{coquad}
\alias{prlin}
\alias{prquad}
\alias{probs}
\alias{linmodoptdesigns}
\alias{truelinmodoptdesigns}
\alias{comp18badoptdesign}
\alias{comp18badoptdesign_got}
\alias{comp18badoptdesign_atk}
\alias{comp15badoptdesign}
\alias{comp15sigoptdesign}
\alias{comp15sigDRSoptdesign}
\alias{LRSIGinidesign}
\alias{LRBADoptdesign}
\alias{LRBADoptdesign2}
\alias{LRSIGoptdesign}
\alias{LRNSELinidesign}
\alias{LRBAAoptdesign}
\alias{LRNSELoptdesign}
\alias{HLRBADoptdesign}
\alias{HLRSIGoptdesign}
\alias{HLRBAAoptdesign}
\alias{HLRNSELoptdesign}

\alias{beetleoptdesign}
\title{
Internal Objects for \code{acebayes}
}
\description{
These are internal objects for the \code{\link{acebayes}} package.
}
\keyword{internal}